робочий js
